# openapi/internal.yaml
openapi: 3.0.3
info:
  title: DC Inventory internal API
  version: 0.0.0
...
paths:
...
  /internal/purchase-orders:
    get:
      operationId: listInternalPurchaseOrders
      summary: List purchase orders
      tags:
        - internal
      parameters:
        - schema:
            type: string
          in: query
          name: q
          required: false
        - schema:
            default: 1
            type: integer
            minimum: 1
            maximum: 9007199254740991
          in: query
          name: page
          required: false
        - schema:
            default: 25
            type: integer
            minimum: 1
            maximum: 100
          in: query
          name: pageSize
          required: false
        - schema:
            default: documentNumber
            type: string
            enum:
              - documentNumber
              - status
          in: query
          name: sortBy
          required: false
        - schema:
            default: asc
            type: string
            enum:
              - asc
              - desc
          in: query
          name: sortOrder
          required: false
        - schema:
            type: string
            enum:
              - draft
              - confirmed
              - received
              - cancelled
          in: query
          name: status
          required: false
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: query
          name: supplierId
          required: false
      x-table:
        rowId: id
        columns:
          - field: documentNumber
            label: "PO #"
          - field: status
            label: Status
          - field: supplierName
            label: Supplier
        search:
          param: q
          fields:
            - documentNumber
          placeholder: Search PO number
        filters:
          - param: status
            control: select
          - param: supplierId
            control: text
        sort:
          defaultBy: documentNumber
          defaultOrder: asc
          fields:
            - documentNumber
            - status
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                        supplierId:
                          type: string
                          format: uuid
                          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                        documentNumber:
                          type: string
                        status:
                          type: string
                          enum:
                            - draft
                            - confirmed
                            - received
                            - cancelled
                        shipDate:
                          nullable: true
                          type: string
                        cancelDate:
                          nullable: true
                          type: string
                        lines:
                          type: array
                          items:
                            type: object
                            properties:
                              id:
                                type: string
                                format: uuid
                                pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                              sku:
                                type: string
                              name:
                                type: string
                              qty:
                                type: integer
                                minimum: -9007199254740991
                                maximum: 9007199254740991
                              receivedQty:
                                type: integer
                                minimum: -9007199254740991
                                maximum: 9007199254740991
                            required:
                              - id
                              - sku
                              - name
                              - qty
                              - receivedQty
                            additionalProperties: false
                        supplierName:
                          type: string
                      required:
                        - id
                        - supplierId
                        - documentNumber
                        - status
                        - shipDate
                        - cancelDate
                        - lines
                        - supplierName
                      additionalProperties: false
                  page:
                    type: integer
                    minimum: -9007199254740991
                    maximum: 9007199254740991
                  pageSize:
                    type: integer
                    minimum: -9007199254740991
                    maximum: 9007199254740991
                  total:
                    type: integer
                    minimum: -9007199254740991
                    maximum: 9007199254740991
                required:
                  - items
                  - page
                  - pageSize
                  - total
                additionalProperties: false
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - invalid_request
                  message:
                    type: string
                    enum:
                      - The request is invalid.
                  requestId:
                    type: string
                required:
                  - error
                  - message
                  - requestId
                additionalProperties: false
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - unauthorized
                required:
                  - error
                additionalProperties: false
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - feature_disabled
                required:
                  - error
                additionalProperties: false
    post:
      operationId: createInternalPurchaseOrder
      summary: Create draft purchase order
      tags:
        - internal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                supplierId:
                  type: string
                  format: uuid
                  pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                shipDate:
                  nullable: true
                  type: string
                  pattern: ^\d{4}-\d{2}-\d{2}$
                cancelDate:
                  nullable: true
                  type: string
                  pattern: ^\d{4}-\d{2}-\d{2}$
                lines:
                  minItems: 1
                  type: array
                  items:
                    type: object
                    properties:
                      sku:
                        type: string
                        minLength: 1
                      name:
                        type: string
                        minLength: 1
                      qty:
                        type: integer
                        minimum: 0
                        exclusiveMinimum: true
                        maximum: 9007199254740991
                    required:
                      - sku
                      - name
                      - qty
              required:
                - supplierId
                - lines
      responses:
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                    pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                  supplierId:
                    type: string
                    format: uuid
                    pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                  documentNumber:
                    type: string
                  status:
                    type: string
                    enum:
                      - draft
                      - confirmed
                      - received
                      - cancelled
                  shipDate:
                    nullable: true
                    type: string
                  cancelDate:
                    nullable: true
                    type: string
                  lines:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                        sku:
                          type: string
                        name:
                          type: string
                        qty:
                          type: integer
                          minimum: -9007199254740991
                          maximum: 9007199254740991
                        receivedQty:
                          type: integer
                          minimum: -9007199254740991
                          maximum: 9007199254740991
                      required:
                        - id
                        - sku
                        - name
                        - qty
                        - receivedQty
                      additionalProperties: false
                required:
                  - id
                  - supplierId
                  - documentNumber
                  - status
                  - shipDate
                  - cancelDate
                  - lines
                additionalProperties: false
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                anyOf:
                  - type: object
                    properties:
                      error:
                        type: string
                        enum:
                          - invalid
                    required:
                      - error
                    additionalProperties: false
                  - type: object
                    properties:
                      error:
                        type: string
                        enum:
                          - invalid_request
                      message:
                        type: string
                        enum:
                          - The request is invalid.
                      requestId:
                        type: string
                    required:
                      - error
                      - message
                      - requestId
                    additionalProperties: false
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - unauthorized
                required:
                  - error
                additionalProperties: false
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                anyOf:
                  - type: object
                    properties:
                      error:
                        type: string
                        enum:
                          - forbidden
                    required:
                      - error
                    additionalProperties: false
                  - type: object
                    properties:
                      error:
                        type: string
                        enum:
                          - feature_disabled
                    required:
                      - error
                    additionalProperties: false
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - not_found
                required:
                  - error
                additionalProperties: false
...
  /internal/purchase-orders/{id}/export:
    get:
      operationId: exportInternalPurchaseOrder
      summary: Export purchase order lines as spreadsheet
      tags:
        - internal
      parameters:
        - schema:
            default: xlsx
            type: string
            enum:
              - xlsx
              - csv
          in: query
          name: format
          required: false
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Spreadsheet file
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
            text/csv:
              schema:
                type: string
                format: binary
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - unauthorized
                required:
                  - error
                additionalProperties: false
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - feature_disabled
                required:
                  - error
                additionalProperties: false
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - not_found
                required:
                  - error
                additionalProperties: false
...
servers:
  - url: http://localhost:3001